+++ src/ui/CursorNode.cpp
@@ -360,8 +360,9 @@ namespace mpedit {
             
             // Draw a bounding box
             auto rect = obj->boundingBox();
-            cocos2d::ccColor4F fill = {color3.r / 255.f, color3.g / 255.f, color3.b / 255.f, 0.2f};
-            cocos2d::ccColor4F border = {color3.r / 255.f, color3.g / 255.f, color3.b / 255.f, 0.8f};
+            int64_t opacityPct = geode::Mod::get()->getSettingValue<int64_t>("selection-overlay-opacity");
+            float alpha = std::clamp(static_cast<float>(opacityPct) / 100.f, 0.f, 1.f);
+            cocos2d::ccColor4F color4 = {color3.r / 255.f, color3.g / 255.f, color3.b / 255.f, alpha};
             
             cocos2d::CCPoint bl = {rect.getMinX(), rect.getMinY()};
             cocos2d::CCPoint tl = {rect.getMinX(), rect.getMaxY()};
@@ -369,7 +370,7 @@ namespace mpedit {
             cocos2d::CCPoint br = {rect.getMaxX(), rect.getMinY()};
             
             cocos2d::CCPoint verts[] = {bl, tl, tr, br};
-            m_selectionDrawNode->drawPolygon(verts, 4, fill, 1.f, border);
+            m_selectionDrawNode->drawPolygon(verts, 4, color4, 1.f, color4);
         }
 
         // Remove disconnected players